update (fe): extracted module item drawer and update editor state
Extracted the add module item drawer logic into a new AddModuleItemDrawer component and replaced AddSubsectionDrawer usage in module-panel.tsx. Updated useEditorState to fetch and provide section data. Cleaned up unused props and improved code organization in CMS and module panel components.
